+++ packages/tolk/test/unit/engine/test_schedule.ml
@@ -61,11 +61,13 @@ let test_allocator =
     Device.Allocator.
       {
         kind = buffer_kind;
+        host = Fun.const None;
+        mapping = None;
+        synchronize = (fun () -> ());
         alloc;
         free;
         copyin;
         copyout;
-        as_buffer = None;
         addr = Some addr;
         offset = None;
         transfer = None;