+++ packages/tolk/test/unit/frontend/test_block_matmul.ml
@@ -186,7 +186,6 @@ let render_only prefix ren =
           free = (fun _ _ _ -> ());
           copyin = (fun _ _ -> ());
           copyout = (fun _ _ -> ());
-          as_buffer = None;
           addr = (fun _ -> Nativeint.zero);
           offset = None;
           transfer = None;